WinGui: Fix an Integer overflow in the picture settings panel.
git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@4916 b64f7644-9d1e-0410-96f1-a4d463321fa5

+++ win/CS/HandBrakeWPF/ViewModels/PictureSettingsViewModel.cs
@@ -1064,7 +1064,7 @@ namespace HandBrakeWPF.ViewModels
 
                         double newHeight = ((double)this.Width * this.sourceResolution.Width * this.SourceAspect.Height *
                                             crop_height) /
-                                           (this.sourceResolution.Height * this.SourceAspect.Width * crop_width);
+                                           ((double)this.sourceResolution.Height * this.SourceAspect.Width * crop_width);
 
                         this.Task.Height = (int)Math.Round(this.GetModulusValue(newHeight), 0);
                         this.NotifyOfPropertyChange(() => this.Height);
